added python catocrypt files

+++ services/src/catocrypt/base64.cc
@@ -72,7 +72,7 @@ encodeBase64(unsigned char const *bytes_to_encode, unsigned int in_len) {
 
 string 
 decodeBase64(string const &encoded_string) {
-    int in_len = encoded_string.size();
+    long in_len = encoded_string.size();
     int i = 0;
     int j = 0;
     int in_ = 0;
